requirements.txt
# --- Cœur de l'application (Flask) ---
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.49
python-dotenv==1.2.2
Werkzeug==3.1.8

# --- Interface Admin & Data (Streamlit) ---
streamlit==1.56.0
streamlit-option-menu==0.3.13
pandas==2.3.3
numpy==1.26.4
openpyxl==3.1.5
altair==6.0.0

# --- Génération QR & Images ---
qrcode==8.2
pillow==12.2.0

# --- Utilitaires ---
requests==2.33.1
python-dateutil==2.9.0.post0
click==8.3.2